+++ src/test/groovy/edu/ie3/datamodel/io/factory/input/MeasurementUnitInputFactoryTest.groovy
@@ -5,11 +5,12 @@
  */
 package edu.ie3.datamodel.io.factory.input
 
-
+import edu.ie3.datamodel.exceptions.FactoryException
 import edu.ie3.datamodel.models.OperationTime
 import edu.ie3.datamodel.models.input.MeasurementUnitInput
 import edu.ie3.datamodel.models.input.NodeInput
 import edu.ie3.datamodel.models.input.OperatorInput
+import edu.ie3.datamodel.utils.Try
 import edu.ie3.test.helper.FactoryTestHelper
 import spock.lang.Specification
 
@@ -38,12 +39,12 @@ class MeasurementUnitInputFactoryTest extends Specification implements FactoryTe
     def nodeInput = Mock(NodeInput)
 
     when:
-    Optional<MeasurementUnitInput> input = inputFactory.get(new NodeAssetInputEntityData(parameter, inputClass, nodeInput))
+    Try<MeasurementUnitInput, FactoryException> input = inputFactory.get(new NodeAssetInputEntityData(parameter, inputClass, nodeInput))
 
     then:
-    input.present
-    input.get().getClass() == inputClass
-    ((MeasurementUnitInput) input.get()).with {
+    input.success
+    input.data().getClass() == inputClass
+    ((MeasurementUnitInput) input.data()).with {
       assert uuid == UUID.fromString(parameter["uuid"])
       assert operationTime == OperationTime.notLimited()
       assert operator == OperatorInput.NO_OPERATOR_ASSIGNED